pyproject.toml
[tool.poetry]
name = "llm_mastodon_agent"
version = "0.1.0"
description = "This repository contains a scalable framework for deploying AI-driven Mastodon accounts using Large Language Models, enabling autonomous social media interactions."
license = "MIT"
authors = ["Simon Münker <simon.muenker@googlemail.de>"]
readme = "README.md"
repository = "https://github.com/simon-muenker/LLM-Mastodon-Agent"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pydantic = "^2.8.2"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.2"
lxml-html-clean = "^0.1.1"
ollama = "^0.2.1"
transformers = "^4.42.4"
torch = "^2.3.1"
accelerate = "^0.32.1"
gnews = "^0.3.7"
newspaper3k = "^0.2.8"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.1"
mypy = "^1.10.1"
pytest = "^8.2.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.NvidiaGPU.dependencies]
bitsandbytes = "^0.43.1"
peft = "^0.11.1"
setuptools = "^71.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 96

[tool.ruff.lint]
select = ["I", "N", "UP", "A", "PL"]


[tool.mypy]
files = "llm_mastodon_agent"
disable_error_code = "import-untyped"


[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
testpaths = ["tests"]
